Refactor to reuse canonicalize
