+++ plotting/plot_compilation_size.py
@@ -1,35 +1,24 @@
 from argparse import ArgumentParser
+
 import matplotlib.pyplot as plt
 import seaborn as sns
-from utils import load_log, find_span
+from utils import load_compilation_logs
 
 argument_parser = ArgumentParser("Stress Test Plotter")
-argument_parser.add_argument("native_logs_path")
+argument_parser.add_argument("logs_path")
 arguments = argument_parser.parse_args()
 
 
-def canonicalize(event):
-    if "native contract compilation finished" not in event["fields"]["message"]:
-        return None
-
-    compilation_span = find_span(event, "contract compilation")
-    if compilation_span is None:
-        return None
-
-    return {
-        "class hash": compilation_span["class_hash"],
-        "size": event["fields"]["size"] / (1024 * 1024),
-    }
-
-
-dataset = load_log(arguments.native_logs_path, canonicalize)
+dataset = load_compilation_logs(
+    arguments.logs_path,
+)
 
 figure, ax = plt.subplots()
 
 sns.set_color_codes("bright")
-sns.violinplot(ax=ax, x="size", data=dataset, cut=0)
+sns.violinplot(ax=ax, x="size", data=dataset[dataset["executor"] == "native"], cut=0)
 
-ax.set_xlabel("Library Size (MiB)")
+ax.set_xlabel("Library Size (KiB)")
 ax.set_title("Library Size by Contract")
 
 plt.show()
