+++ plotting/plot_compilation_size_corr.py
@@ -2,36 +2,14 @@ from argparse import ArgumentParser
 
 import matplotlib.pyplot as plt
 import seaborn as sns
-from utils import load_log, find_span
+from utils import load_compilation_logs
 
 argument_parser = ArgumentParser("Stress Test Plotter")
 argument_parser.add_argument("logs_path")
 arguments = argument_parser.parse_args()
 
 
-def canonicalize(event):
-    if "contract compilation finished" not in event["fields"]["message"]:
-        return None
-
-    compilation_span = find_span(event, "contract compilation")
-    if compilation_span is None:
-        return None
-
-    if "vm" in event["fields"]["message"]:
-        executor = "vm"
-    elif "native" in event["fields"]["message"]:
-        executor = "native"
-    else:
-        raise Exception("Invalid Executor")
-
-    return {
-        "class hash": compilation_span["class_hash"],
-        "size": event["fields"]["size"] / 1024,
-        "executor": executor,
-    }
-
-
-dataset = load_log(arguments.logs_path, canonicalize)
+dataset = load_compilation_logs(arguments.logs_path)
 dataset = dataset.pivot_table(index="class hash", columns="executor")
 dataset.columns = ["_".join(a) for a in dataset.columns.to_flat_index()]
