+++ plotting/utils.py
@@ -12,10 +12,36 @@ def find_span(event, name):
     return None
 
 
-CHUNKSIZE = 100000
+def load_compilation_logs(path):
+    def canonicalize(event):
+        # keep contract compilation finished logs
+        if "contract compilation finished" not in event["fields"]["message"]:
+            return None
+
+        compilation_span = find_span(event, "contract compilation")
+        if compilation_span is None:
+            return None
+
+        if "vm" in event["fields"]["message"]:
+            executor = "vm"
+        elif "native" in event["fields"]["message"]:
+            executor = "native"
+        else:
+            raise Exception("Invalid Executor")
+
+        return {
+            "class hash": compilation_span["class_hash"],
+            "time": float(event["fields"]["time"]),
+            "size": float(event["fields"]["size"]) / 1024,
+            "length": float(compilation_span["length"]) / 1024,
+            "executor": executor,
+        }
+
+    return load_jsonl(path, canonicalize)
 
 
-def load_log(path, f):
+def load_jsonl(path, f):
+    CHUNKSIZE = 100000
     dataset = pd.DataFrame()
 
     with pd.read_json(path, lines=True, typ="series", chunksize=CHUNKSIZE) as chunks:
